fix: objectId 자료형 통일, 서버 메모리 누수 해결, 퇴장 시 Tank 제거로직 추가.

+++ Sentinel/Sentinel/GameObject.cpp
@@ -5,7 +5,7 @@ GameObject::GameObject()
 {
 }
 
-GameObject::GameObject(UINT32 id)
+GameObject::GameObject(UINT16 id)
 	: _transform{ 0.f, 0.f, 0.f, 1.f, 0.f, 0.f, 0.f }
 	, _id(id)
 {
@@ -15,7 +15,7 @@ GameObject::~GameObject()
 {
 }
 
-UINT32 GameObject::GetID() const
+UINT16 GameObject::GetID() const
 {
 	return _id;
 }
